ml_services/requirements.txt
Flask
requests
PyPDF2
python-dotenv

